+++ CHANGELOG.md
@@ -24,3 +24,23 @@ the following form: YYYY.0M.0D.
 
 - Changed design to include changeElement(), getElement(), rows(), columns(), elementAdd(), and dotProduct.
 
+## 2025.03.24
+
+### Added
+
+- Designed abstract class for Matrix component
+
+### Updated
+
+- Changed design to include the implementation of the secondary methods. No new methods and no methods removed from original plan. All methods were able to be implemented, and no new need for any more methods (yet).
+
+## 2025.03.03
+
+### Added Interfaces
+
+- Designed kernel and enhanced interfaces for Matrix component
+
+### Updated
+
+- Changed design to include the kernel interfaace and the enhanced interface.
+